Improved the package script in Linux.

+++ utils/scripts/package_ddsrecorder.sh
@@ -12,16 +12,13 @@
 
 errorstatus=0
 
+# Default value, oll platforms will be generated (i86, x64)
+package_targets=""
+
 function package
 {
     # Get current version of GCC.
     . $EPROSIMADIR/scripts/common_pack_functions.sh getGccVersion
-
-    # Get the initials of the target.
-    . $EPROSIMADIR/scripts/common_pack_functions.sh getTargetFromEprosimaTarget
-
-    #Change EPROSIMA_TARGET
-    EPROSIMA_TARGET="${etarget}Linux2.6gcc${gccversion}"
 
     # Update and compile CDR library.
     cd ../CDR
@@ -29,12 +26,26 @@ function package
     svn update
     errorstatus=$?
     if [ $errorstatus != 0 ]; then return; fi
-    # Compile CDR library for target.
-    rm -rf output
-    rm -r lib/$EPROSIMA_TARGET
-    make
-    errorstatus=$?
-    if [ $errorstatus != 0 ]; then return; fi
+    # Compile CDR library for target i86.
+    if [ -z $package_targets ] || [ "$package_targets" == "i86" ]; then
+        #Change EPROSIMA_TARGET
+        EPROSIMA_TARGET="i86Linux2.6gcc${gccversion}"
+        rm -rf output
+        rm -r lib/$EPROSIMA_TARGET
+        make
+        errorstatus=$?
+        if [ $errorstatus != 0 ]; then return; fi
+    fi
+    # Compile CDR library for target x64.
+    if [ -z $package_targets ] || [ "$package_targets" == "x64" ]; then
+        #Change EPROSIMA_TARGET
+        EPROSIMA_TARGET="x64Linux2.6gcc${gccversion}"
+        rm -rf output
+        rm -r lib/$EPROSIMA_TARGET
+        make
+        errorstatus=$?
+        if [ $errorstatus != 0 ]; then return; fi
+    fi
     cd ../DDSRecorder
 
     # Get the current version of DDSRecorder
@@ -47,11 +58,25 @@ function package
     svn update
     errorstatus=$?
     if [ $errorstatus != 0 ]; then return; fi
-    # Compile DDSRecorder for target.
-    rm -rf output
-    make
-    errorstatus=$?
-    if [ $errorstatus != 0 ]; then return; fi
+    # Compile DDSRecorder for target i86.
+    if [ -z $package_targets ] || [ "$package_targets" == "i86" ]; then
+        #Change EPROSIMA_TARGET
+        EPROSIMA_TARGET="i86Linux2.6gcc${gccversion}"
+        # Compile DDSRecorder for target.
+        rm -rf output
+        make
+        errorstatus=$?
+        if [ $errorstatus != 0 ]; then return; fi
+    fi
+    if [ -z $package_targets ] || [ "$package_targets" == "x64" ]; then
+        #Change EPROSIMA_TARGET
+        EPROSIMA_TARGET="x64Linux2.6gcc${gccversion}"
+        # Compile DDSRecorder for target.
+        rm -rf output
+        make
+        errorstatus=$?
+        if [ $errorstatus != 0 ]; then return; fi
+    fi
 
     # Create PDFS from documentation.
     cd doc
@@ -76,9 +101,22 @@ function package
 
     # Create installers
     cd utils/installers/linux
-    ./setup_linux.sh $recorderversion
-    errorstatus=$?
-    if [ $errorstatus != 0 ]; then return; fi
+    #~Installer for target i86.
+    if [ -z $package_targets ] || [ "$package_targets" == "i86" ]; then
+        #Change EPROSIMA_TARGET
+        EPROSIMA_TARGET="i86Linux2.6gcc${gccversion}"
+        ./setup_linux.sh $recorderversion
+        errorstatus=$?
+        if [ $errorstatus != 0 ]; then return; fi
+    fi
+    #~Installer for target x64
+    if [ -z $package_targets ] || [ "$package_targets" == "x64" ]; then
+        #Change EPROSIMA_TARGET
+        EPROSIMA_TARGET="x64Linux2.6gcc${gccversion}"
+        ./setup_linux.sh $recorderversion
+        errorstatus=$?
+        if [ $errorstatus != 0 ]; then return; fi
+    fi
     cd ../../..
 }
 
@@ -86,6 +124,16 @@ function package
 if [ "$EPROSIMADIR" == "" ]; then
     echo "environment.sh must to be run."
     exit -1
+fi
+
+# Get the optional parameter
+if [ $# -ge 1 ] && [ -n $1 ]; then
+    if [ "$1" == "i86" ] || [ "$1" == "x64" ]; then
+        package_targets=$1
+    else
+        echo Error: Bad argument. Valid values. i86, x64
+        exit -1;
+    fi
 fi
 
 # Go to root directory
